waitForExist method added to allerMessage, examples of webdriverio deleted, README modifed and rename of page.js to url_page.js
