--- test/pageobjects/example/secure.page.js
@@ -1,17 +1,0 @@
-
-
-import Page from './page.js';
-
-/**
- * sub page containing specific selectors and methods for a specific page
- */
-class SecurePage extends Page {
-    /**
-     * define selectors using getter methods
-     */
-    get flashAlert () {
-        return $('#flash');
-    }
-}
-
-export default new SecurePage();
