--- test/specs/example.e2e.js
@@ -1,14 +1,0 @@
-import LoginPage from '../pageobjects/example/login.page.js';
-import SecurePage from '../pageobjects/example/secure.page.js';
-
-describe('My Login application', () => {
-  it('should login with valid credentials', async () => {
-    await LoginPage.open();
-
-    await LoginPage.login('tomsmith', 'SuperSecretPassword!');
-    await expect(SecurePage.flashAlert).toBeExisting();
-    await expect(SecurePage.flashAlert).toHaveTextContaining(
-      'You logged into a secure area!'
-    );
-  });
-});
